package names to spec
