--- core/src/com/mygdx/insanerpg/Item/Equipment/Armor/Armor.java
@@ -1,6 +1,0 @@
-package com.mygdx.insanerpg.Item.Equipment.Armor;
-
-import com.mygdx.insanerpg.Item.Equipment.Equipment;
-
-public class Armor extends Equipment {
-}
