--- core/src/com/mygdx/insanerpg/Item/Equipment/Equipment.java
@@ -1,7 +1,0 @@
-package com.mygdx.insanerpg.Item.Equipment;
-
-import com.mygdx.insanerpg.Item.Item;
-
-public class Equipment extends Item {
-    private int _durability;
-}
