--- core/src/com/mygdx/insanerpg/Item/Equipment/Weapon/Scythe.java
@@ -1,4 +1,0 @@
-package com.mygdx.insanerpg.Item.Equipment.Weapon;
-
-public class Scythe extends Weapon{
-}
