--- core/src/com/mygdx/insanerpg/Item/Equipment/Weapon/Sword.java
@@ -1,6 +1,0 @@
-package com.mygdx.insanerpg.Item.Equipment.Weapon;
-
-
-public class Sword extends Weapon{
-
-}
