--- core/src/com/mygdx/insanerpg/Item/Stackable/Consumable/Potion.java
@@ -1,5 +1,0 @@
-package com.mygdx.insanerpg.Item.Stackable.Consumable;
-
-public class Potion extends Consumable {
-
-}
