--- core/src/com/mygdx/insanerpg/Item/Stackable/Stackable.java
@@ -1,6 +1,0 @@
-package com.mygdx.insanerpg.Item.Stackable;
-
-import com.mygdx.insanerpg.Item.Item;
-
-public class Stackable extends Item {
-}
