+++ core/src/com/mygdx/insanerpg/item/equipment/weapon/Scythe.java
@@ -1,0 +1,4 @@
+package com.mygdx.insanerpg.item.equipment.weapon;
+
+public class Scythe extends Weapon{
+}
