+++ core/src/com/mygdx/insanerpg/item/equipment/weapon/Sword.java
@@ -1,0 +1,6 @@
+package com.mygdx.insanerpg.item.equipment.weapon;
+
+
+public class Sword extends Weapon{
+
+}
