+++ core/src/com/mygdx/insanerpg/item/stackable/Stackable.java
@@ -1,0 +1,6 @@
+package com.mygdx.insanerpg.item.stackable;
+
+import com.mygdx.insanerpg.item.Item;
+
+public class Stackable extends Item {
+}
