+++ core/src/com/mygdx/insanerpg/item/stackable/consumable/Consumable.java
@@ -1,6 +1,6 @@
-package com.mygdx.insanerpg.Item.Stackable.Consumable;
+package com.mygdx.insanerpg.item.stackable.consumable;
 
-import com.mygdx.insanerpg.Item.Item;
+import com.mygdx.insanerpg.item.Item;
 
 public class Consumable extends Item {
     private int _cooldown;
